requirements.txt
asgiref==3.3.1
autopep8==1.5.5
Django==3.1.3
django-cors-headers==3.7.0
django-crispy-forms==1.10.0
django-filter==2.4.0
djangorestframework==3.12.2
Pillow==8.0.1
pycodestyle==2.6.0
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
